Try explicitly setting the SSH port
I'm looking at the connection traffic on the server to see why it's failing, and I think maybe this Action is just trying to connect with a random port each time. Of course that makes no sense because the default port is supposed to be 22, but I'm running out of ideas right now. Why isn't this working