Remove mention of events
We're taking out the wording about events until we have our event schedule made for this semester.

Main: removed the wording about the weekly meetings on the home page, but kept the Discord invite.
Events: removed the weekly events card and left just the "Coming Soon" message and the past highlights.