Add attendee limit
If the event is at its max number of registrations, the website will not allow you to fill out the registration form, and the server will reject any further registration fee payments.

I've hacked this in at the moment with a solution I don't particularly like but is easy to implement. Once I've figured out how to do this with the Google Sheets API, I'll replace this implementation.