Convert Google Form to traditional form
Embedding a Google Form onto the site turned out to be problematic, because the form redirects to accounts.google.com to log in when you aren't signed into Google in your browser, and accounts.google.com does not allow itself to be embedded. It just seems easier consolidating the Google part of the form with the traditional part of the form.