React to one attendance type being full
If one attendance type is full, it will be grayed out and the customer will only be able to choose the other option. They will also see a notice on the page about this.

Fun fact! GitHub Copilot wrote almost all of this! I only gave it the `else if (event.full.inPerson)` line.